+++ packages/router-plugin/src/router.state.ts
@@ -5,9 +5,14 @@ import {
   Router,
   RouterStateSnapshot,
   RoutesRecognized,
-  ResolveEnd
+  ResolveEnd,
+  NavigationEnd,
+  GuardsCheckEnd
 } from '@angular/router';
+import { Location } from '@angular/common';
 import { Action, Selector, State, StateContext, Store } from '@ngxs/store';
+import { isAngularInTestMode } from '@ngxs/store/internals';
+import { filter, take } from 'rxjs/operators';
 
 import {
   Navigate,
@@ -37,6 +42,7 @@ export class RouterState {
   private lastRoutesRecognized: RoutesRecognized;
   private dispatchTriggeredByRouter = false; // used only in dev mode in combination with routerReducer
   private navigationTriggeredByDispatch = false; // used only in dev mode in combination with routerReducer
+  private lastResolvedStateSnapshot: RouterStateSnapshot = null!;
 
   /**
    * Selectors
@@ -56,10 +62,12 @@ export class RouterState {
     private _store: Store,
     private _router: Router,
     private _serializer: RouterStateSerializer<RouterStateSnapshot>,
-    private _ngZone: NgZone
+    private _ngZone: NgZone,
+    private location: Location
   ) {
     this.setUpStoreListener();
     this.setUpStateRollbackEvents();
+    this.checkInitialNavigationOnce();
   }
 
   @Action(Navigate)
@@ -97,8 +105,11 @@ export class RouterState {
     this._router.events.subscribe(e => {
       if (e instanceof RoutesRecognized) {
         this.lastRoutesRecognized = e;
-      } else if (e instanceof ResolveEnd) {
-        this.resolveEnd(e.state);
+      } else if (e instanceof GuardsCheckEnd || e instanceof ResolveEnd) {
+        // The `GuardsCheckEnd` event is always triggered unlike the `ResolveEnd`
+        this.lastResolvedStateSnapshot = e.state;
+      } else if (e instanceof NavigationEnd) {
+        this.navigationEnd();
       } else if (e instanceof NavigationCancel) {
         this.dispatchRouterCancel(e);
       } else if (e instanceof NavigationError) {
@@ -107,12 +118,8 @@ export class RouterState {
     });
   }
 
-  /**
-   * The `ResolveEnd` event is always triggered after running all resolvers
-   * that are linked to some route and child routes
-   */
-  private resolveEnd(routerStateSnapshot: RouterStateSnapshot): void {
-    this.routerStateSnapshot = this._serializer.serialize(routerStateSnapshot);
+  private navigationEnd(): void {
+    this.routerStateSnapshot = this._serializer.serialize(this.lastResolvedStateSnapshot);
     if (this.shouldDispatchRouterNavigation()) {
       this.dispatchRouterNavigation();
     }
@@ -124,10 +131,9 @@ export class RouterState {
   }
 
   private navigateIfNeeded(): void {
-    if (!this.routerState || !this.routerState.state) {
+    if (!this.routerState || !this.routerState.state || this.dispatchTriggeredByRouter) {
       return;
     }
-    if (this.dispatchTriggeredByRouter) return;
 
     if (this._router.url !== this.routerState.state.url) {
       this.navigationTriggeredByDispatch = true;
@@ -173,5 +179,33 @@ export class RouterState {
       this.dispatchTriggeredByRouter = false;
       this.navigationTriggeredByDispatch = false;
     }
+  }
+
+  /**
+   * No sense to mess up the `setUpStateRollbackEvents` method as we have
+   * to perform this check only once and unsubscribe after the first event
+   * is triggered
+   */
+  private checkInitialNavigationOnce(): void {
+    if (isAngularInTestMode()) {
+      return;
+    }
+
+    this._router.events
+      .pipe(
+        filter((event): event is RoutesRecognized => event instanceof RoutesRecognized),
+        take(1)
+      )
+      .subscribe(({ url }) => {
+        // `location.pathname` always equals manually entered URL in the address bar
+        // e.g. `location.pathname === '/foo'`, but the `router` state has been initialized
+        // with another URL (e.g. used in combination with `NgxsStoragePlugin`), thus the
+        // `RouterNavigation` action will be dispatched and the user will be redirected to the
+        // previously saved URL. We want to prevent such behavior, so we perform this check
+        // in order to redirect user to the manually entered URL if it differs from the recognized one
+        if (url !== this.location.path()) {
+          this._router.navigateByUrl(location.pathname);
+        }
+      });
   }
 }