+++ packages/store/internals/src/index.ts
@@ -1,1 +1,3 @@
+export { isAngularInTestMode } from './angular';
 export { NgxsBootstrapper } from './ngxs-bootstrapper';
+export { memoize } from './memoize';