+++ packages/store/src/module.ts
@@ -1,113 +1,55 @@
 import {
-  NgModule,
-  ModuleWithProviders,
-  Optional,
-  Inject,
+  APP_BOOTSTRAP_LISTENER,
   InjectionToken,
-  APP_BOOTSTRAP_LISTENER
+  isDevMode,
+  ModuleWithProviders,
+  NgModule,
+  Provider
 } from '@angular/core';
+import { isAngularInTestMode, NgxsBootstrapper } from '@ngxs/store/internals';
 
-import { NgxsBootstrapper } from '@ngxs/store/internals';
-
-import { ROOT_STATE_TOKEN, FEATURE_STATE_TOKEN, NgxsConfig } from './symbols';
+import {
+  FEATURE_STATE_TOKEN,
+  NG_TEST_MODE,
+  NG_DEV_MODE,
+  NgxsConfig,
+  ROOT_STATE_TOKEN
+} from './symbols';
 import { NGXS_EXECUTION_STRATEGY } from './execution/symbols';
 import { StateFactory } from './internal/state-factory';
 import { StateContextFactory } from './internal/state-context-factory';
 import { Actions, InternalActions } from './actions-stream';
-import { ConfigValidator } from './internal/config-validator';
 import { LifecycleStateManager } from './internal/lifecycle-state-manager';
-import { InternalDispatcher, InternalDispatchedActionResults } from './internal/dispatcher';
+import { InternalDispatchedActionResults, InternalDispatcher } from './internal/dispatcher';
 import { InternalStateOperations } from './internal/state-operations';
 import { Store } from './store';
 import { SelectFactory } from './decorators/select';
 import { StateStream } from './internal/state-stream';
 import { PluginManager } from './plugin-manager';
-import { InitState, UpdateState } from './actions/actions';
 import { StateClass } from './internal/internals';
+import { NgxsRootModule } from './modules/ngxs-root.module';
+import { NgxsFeatureModule } from './modules/ngxs-feature.module';
 import { DispatchOutsideZoneNgxsExecutionStrategy } from './execution/dispatch-outside-zone-ngxs-execution-strategy';
 import { InternalNgxsExecutionStrategy } from './execution/internal-ngxs-execution-strategy';
+import { HostEnvironment } from './host-environment/host-environment';
+import { ConfigValidator } from './internal/config-validator';
 
-/**
- * Root module
- * @ignore
- */
-@NgModule()
-export class NgxsRootModule {
-  constructor(
-    factory: StateFactory,
-    internalStateOperations: InternalStateOperations,
-    store: Store,
-    select: SelectFactory,
-    @Optional()
-    @Inject(ROOT_STATE_TOKEN)
-    states: StateClass[] = [],
-    lifecycleStateManager: LifecycleStateManager
-  ) {
-    // add stores to the state graph and return their defaults
-    const results = factory.addAndReturnDefaults(states);
-
-    internalStateOperations.setStateToTheCurrentWithNew(results);
-
-    // connect our actions stream
-    factory.connectActionHandlers();
-
-    // dispatch the init action and invoke init and bootstrap functions after
-    lifecycleStateManager.ngxsBootstrap(new InitState(), results);
-  }
-}
-
-/**
- * Feature module
- * @ignore
- */
-@NgModule({})
-export class NgxsFeatureModule {
-  constructor(
-    store: Store,
-    internalStateOperations: InternalStateOperations,
-    factory: StateFactory,
-    @Optional()
-    @Inject(FEATURE_STATE_TOKEN)
-    states: any[][],
-    lifecycleStateManager: LifecycleStateManager
-  ) {
-    // Since FEATURE_STATE_TOKEN is a multi token, we need to
-    // flatten it [[Feature1State, Feature2State], [Feature3State]]
-    const flattenedStates = ([] as any[]).concat(...states);
-
-    // add stores to the state graph and return their defaults
-    const results = factory.addAndReturnDefaults(flattenedStates);
-
-    if (results.states.length) {
-      internalStateOperations.setStateToTheCurrentWithNew(results);
-      // dispatch the update action and invoke init and bootstrap functions after
-      lifecycleStateManager.ngxsBootstrap(new UpdateState(), results);
-    }
-  }
-}
-
-export type ModuleOptions = Partial<NgxsConfig>;
-
-export function ngxsConfigFactory(options: ModuleOptions): NgxsConfig {
-  const config = Object.assign(new NgxsConfig(), options);
-  return config;
-}
-
-export function appBootstrapListenerFactory(bootstrapper: NgxsBootstrapper) {
-  return () => bootstrapper.bootstrap();
-}
-
-export const ROOT_OPTIONS = new InjectionToken<ModuleOptions>('ROOT_OPTIONS');
+type NgxsModuleOptions = Partial<NgxsConfig>;
 
 /**
  * Ngxs Module
  */
-@NgModule({})
+@NgModule()
 export class NgxsModule {
+  private static readonly ROOT_OPTIONS = new InjectionToken<NgxsModuleOptions>('ROOT_OPTIONS');
+
   /**
    * Root module factory
    */
-  static forRoot(states: StateClass[] = [], options: ModuleOptions = {}): ModuleWithProviders {
+  public static forRoot(
+    states: StateClass[] = [],
+    options: NgxsModuleOptions = {}
+  ): ModuleWithProviders {
     return {
       ngModule: NgxsRootModule,
       providers: [
@@ -117,6 +59,7 @@ export class NgxsModule {
         InternalActions,
         NgxsBootstrapper,
         ConfigValidator,
+        HostEnvironment,
         LifecycleStateManager,
         InternalDispatcher,
         InternalDispatchedActionResults,
@@ -127,29 +70,7 @@ export class NgxsModule {
         SelectFactory,
         PluginManager,
         ...states,
-        {
-          provide: NGXS_EXECUTION_STRATEGY,
-          useClass: options.executionStrategy || DispatchOutsideZoneNgxsExecutionStrategy
-        },
-        {
-          provide: ROOT_STATE_TOKEN,
-          useValue: states
-        },
-        {
-          provide: ROOT_OPTIONS,
-          useValue: options
-        },
-        {
-          provide: NgxsConfig,
-          useFactory: ngxsConfigFactory,
-          deps: [ROOT_OPTIONS]
-        },
-        {
-          provide: APP_BOOTSTRAP_LISTENER,
-          useFactory: appBootstrapListenerFactory,
-          multi: true,
-          deps: [NgxsBootstrapper]
-        }
+        ...NgxsModule.ngxsTokenProviders(states, options)
       ]
     };
   }
@@ -157,7 +78,7 @@ export class NgxsModule {
   /**
    * Feature module factory
    */
-  static forFeature(states: StateClass[] = []): ModuleWithProviders {
+  public static forFeature(states: StateClass[] = []): ModuleWithProviders {
     return {
       ngModule: NgxsFeatureModule,
       providers: [
@@ -171,5 +92,60 @@ export class NgxsModule {
         }
       ]
     };
+  }
+
+  private static ngxsTokenProviders(
+    states: StateClass[],
+    options: NgxsModuleOptions
+  ): Provider[] {
+    return [
+      {
+        provide: NG_DEV_MODE,
+        useFactory: NgxsModule.isAngularInTestMode
+      },
+      {
+        provide: NG_TEST_MODE,
+        useFactory: NgxsModule.isAngularDevMode
+      },
+      {
+        provide: NGXS_EXECUTION_STRATEGY,
+        useClass: options.executionStrategy || DispatchOutsideZoneNgxsExecutionStrategy
+      },
+      {
+        provide: ROOT_STATE_TOKEN,
+        useValue: states
+      },
+      {
+        provide: NgxsModule.ROOT_OPTIONS,
+        useValue: options
+      },
+      {
+        provide: NgxsConfig,
+        useFactory: NgxsModule.ngxsConfigFactory,
+        deps: [NgxsModule.ROOT_OPTIONS]
+      },
+      {
+        provide: APP_BOOTSTRAP_LISTENER,
+        useFactory: NgxsModule.appBootstrapListenerFactory,
+        multi: true,
+        deps: [NgxsBootstrapper]
+      }
+    ];
+  }
+
+  private static ngxsConfigFactory(options: NgxsModuleOptions): NgxsConfig {
+    return Object.assign(new NgxsConfig(), options);
+  }
+
+  private static appBootstrapListenerFactory(bootstrapper: NgxsBootstrapper): Function {
+    return () => bootstrapper.bootstrap();
+  }
+
+  private static isAngularInTestMode(): Function {
+    return () => isAngularInTestMode();
+  }
+
+  private static isAngularDevMode(): Function {
+    return () => isDevMode();
   }
 }