+++ CHANGELOG.md
@@ -4,9 +4,24 @@
 ```bash
 $ npm install @ngxs/store@dev
 ```
-* ...
+
+* Feature: selector option to disable supressing errors [#1015](https://github.com/ngxs/store/pull/1015)
+* Fix: Logger Plugin - replace `Object.entries` as it breaks IE11 [#931](https://github.com/ngxs/store/pull/931)
+* Fix: Router Plugin - trigger navigation on the `QueryParams` change [#924](https://github.com/ngxs/store/pull/924)
+* Fix: Router-plugin - redirect to the manually entered route [#920](https://github.com/ngxs/store/pull/920)
+* Fix: HMR-Plugin - remove old styles after reload [#1001](https://github.com/ngxs/store/pull/1001)
+
+## NGXS-Labs
+
+### Immer-adapter v3.0.0
+
+* Feature: immutable state context decorator
+* Feature: immutable selector decorator
+* Feature: support immer v3.x
+* Deprecated: produce operator
 
 # 3.4.3 2019-03-14
+* Feature: Add state defaults to UpdateState [#956](https://github.com/ngxs/store/pull/956)
 * Fix: Router Plugin - serialize after Resolvers have run [#895](https://github.com/ngxs/store/pull/895)
 * Fix: HMR Plugin - incorrect destruction of modules in hmr [#908](https://github.com/ngxs/store/pull/908)
 * Fix: Logger Plugin - print action properties [#879](https://github.com/ngxs/store/pull/879)